moved minimap controller classes into seperate files
