+++ EpicLoot/src/Adventure/Minimap/MinimapPinQueueTask.cs
@@ -1,0 +1,10 @@
+﻿namespace EpicLoot;
+
+public enum MinimapPinQueueTask
+{
+    AddTreasurePin,
+    AddBountyPin,
+    RemoveTreasurePin,
+    RemoveBountyPin,
+    RefreshAll
+}
